+++ src/Symfony/Component/Debug/CHANGELOG.md
@@ -5,6 +5,9 @@ CHANGELOG
 -----
 
 * made the `ErrorHandler` and `ExceptionHandler` classes final
+* added `Exception\FlattenException::getAsString` and
+`Exception\FlattenException::getTraceAsString` to increase compatibility to php
+exception objects
 
 4.0.0
 -----